Move all typed dictionaries to the utils file for it.
